feat: implement internationalization setup and update router imports
