--- src/i18n.js
@@ -1,12 +1,0 @@
-import {createI18n} from "vue-i18n";
-import en from "./locales/en.json";
-import es from "./locales/es.json";
-
-const i18n = createI18n({
-    legacy: false,
-    locale: 'en',
-    fallbackLocale: 'en',
-    messages: { en, es }
-});
-
-export default i18n;
