+++ src/router/router.js
@@ -1,8 +1,8 @@
 import { createRouter, createWebHistory } from "vue-router";
-import Home from "./shared/presentation/views/home.vue";
+import Home from "../shared/presentation/views/home.vue";
 
 const pageNotFound = () =>
-  import("./shared/presentation/views/page-not-found.vue");
+  import("../shared/presentation/views/page-not-found.vue");
 const routes = [
   { path: "/home", name: "home", component: Home, meta: { title: "Home" } },
   { path: "/", redirect: "/home" },
